% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CherryPickPalette.R
\name{CherryPickPalette}
\alias{CherryPickPalette}
\title{Cherry Pick Palette --- Getting Facelift}
\usage{
CherryPickPalette(name, name2 = NULL, name3 = NULL, ...)
}
\arguments{
\item{name}{Name of 1st palette}

\item{name2}{Name of 2nd palette}

\item{name3}{Name of 3nd (optional) palette}

\item{...}{Numeric, complex, or logical vectors}
}
\value{
user-defined palette of colors
}
\description{
This function allows user to cherry pick colors from 2 to 3 palettes
}
\examples{
CherryPickPalette("GoldenTemple","AmritsariPedeWaliLassi")
CherryPickPalette("BiryaniRice","Kulfi","Haveli2")
}
