% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{ListPalette}
\alias{ListPalette}
\title{List Palette}
\usage{
ListPalette(listname)
}
\arguments{
\item{listname}{Name of palette vector}
}
\value{
List of palettes
}
\description{
This function returns list of all palettes
}
\examples{
ListPalette("PunjabiPalette")
}
