% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{MergePalette}
\alias{MergePalette}
\title{Merge Palette}
\usage{
MergePalette(name, name2)
}
\arguments{
\item{string, string}{2 palette names}
}
\value{
Vector of 10 color values (in the unlikely event there are duplicates, then return less than 10)
}
\description{
Merge Palette
}
\examples{
MergePalette("Gidha", "Teej")
}
