% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{RanglaPunjab}
\alias{RanglaPunjab}
\title{Palette of 5 Colors}
\usage{
RanglaPunjab(name = NULL, ...)
}
\arguments{
\item{name}{Palette name}

\item{...}{Numeric, complex, or logical vectors}
}
\value{
Vector of 5 color values
}
\description{
This function returns a palette of 5 colors
}
\examples{
RanglaPunjab("GoldenTemple")
RanglaPunjab("SohniMahiwal")
RanglaPunjab("Teej")
}
