% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPalettePhoto.R
\name{ShowPalettePhoto}
\alias{ShowPalettePhoto}
\title{Show Palette Photo --- Getting Facelift}
\usage{
ShowPalettePhoto(name = NULL, ...)
}
\arguments{
\item{name}{palette name}

\item{...}{Numeric, complex, or logical vectors}
}
\value{
image of reference photo with palette
}
\description{
This function shows photo that inspired a palette
}
\examples{
ShowPalettePhoto("GoldenTemple")
ShowPalettePhoto("Kulfi")
ShowPalettePhoto("AmritsariPedeWaliLassi")
}
