% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortPalette.R
\name{SortPalette}
\alias{SortPalette}
\title{Sort Palette - Under Construction https://www.alanzucconi.com/2015/09/30/colour-sorting/}
\usage{
SortPalette(name = NULL, name2 = NULL, name3 = NULL, ...)
}
\arguments{
\item{name}{Name of 1st palette}

\item{name2}{Name of 2nd (optional) palette}

\item{name3}{Name of 3rd (optional) palette}

\item{...}{Numeric, complex, or logical vectors}
}
\value{
Vector of 10 or 15 sorted color values (in the unlikely event there are duplicates, 
        then return less than 10 or 15 colors)
}
\description{
This function sorts colors of up to 3 palettes
}
\examples{
SortPalette("Kulfi")
SortPalette("Haveli","HeerRanjha")
SortPalette("FieldsOfPunjab2","GoldenTemple2","Jutti3")
}
