% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{MergePalette}
\alias{MergePalette}
\title{Merge Palette}
\usage{
MergePalette(name, name2)
}
\arguments{
\item{string, }{string 1 or 2 palette name(s)}
}
\value{
Vector of 10 color values (in the unlikely event there are duplicates, then return less than 10)
}
\description{
This function returns a palette of 10 colors (in the unlikely event there are duplicates, then return less than 10)
}
\examples{
MergePalette("Gidha", "Teej")
}
