% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{PaintPalette}
\alias{PaintPalette}
\title{Paint Palette}
\usage{
PaintPalette(name, name2)
}
\arguments{
\item{name}{Name of 1st palette}

\item{name2}{Name of 2nd (optional) palette}
}
\value{
image of colors and description
}
\description{
This function paints an image of 1 or 2 palettes
}
\examples{
PaintPalette("Pindh")
PaintPalette("FieldsOfPunjab","Jutti")
}
