% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{PaintPalette}
\alias{PaintPalette}
\title{Paint Palette}
\usage{
PaintPalette(name = NULL, name2 = NULL, name3 = NULL, ...)
}
\arguments{
\item{name}{Name of 1st palette}

\item{name2}{Name of 2nd (optional) palette}

\item{name3}{Name of 3rd (optional) palette}

\item{...}{Numeric, complex, or logical vectors}
}
\value{
image of colors and description
}
\description{
This function paints an image of up to 3 palettes
}
\examples{
PaintPalette("Pindh")
PaintPalette("FieldsOfPunjab","Jutti")
PaintPalette("FieldsOfPunjab","Jutti","Paranda")
}
