% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{PaintPalette}
\alias{PaintPalette}
\title{Paint Palette}
\usage{
PaintPalette(name, name2)
}
\arguments{
\item{string, string}{1 or 2 palette names}
}
\value{
image of colors and description
}
\description{
Paint Palette
}
\examples{
PaintPalette("Pindh")
PaintPalette("FieldsOfPunjab","Jutti")
}
