% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{RanglaPunjab}
\alias{RanglaPunjab}
\title{Palette of 5 Colors}
\usage{
RanglaPunjab(name)
}
\arguments{
\item{string}{Palette name}
}
\value{
Vector of 5 color values
}
\description{
Palette of 5 Colors
}
\examples{
RanglaPunjab("Teej")
}
