% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanglaPunjab.R
\name{ShowPalettePhoto}
\alias{ShowPalettePhoto}
\title{Show Palette Photo}
\usage{
ShowPalettePhoto(name)
}
\arguments{
\item{string}{palette name}
}
\value{
image of reference photo with palette
}
\description{
This function shows photo that inspired a palette
}
\examples{
ShowPalettePhoto("GoldenTemple")
}
